packages/agent-core/pyproject.toml
# packages/agent-core/pyproject.toml
# @author Ysf

[project]
name = "agent-core"
version = "0.1.0"
description = "AI Creator Agent 核心共享包"
requires-python = ">=3.11"
dependencies = [
    # 核心依赖
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    # LangChain/LangGraph
    "langchain>=0.1.0",
    "langgraph>=0.2.0",
    # LLM API
    "anthropic>=0.40.0",
    # 表达式解析
    "jsonpath-ng>=1.6.0",
    "simpleeval>=0.9.13",
    # 加密
    "cryptography>=42.0.0",
    # 云存储
    "boto3>=1.34.0",
    # 异步 HTTP
    "aiohttp>=3.9.0",
    # 异步文件操作
    "aiofiles>=23.0.0",
]

[project.optional-dependencies]
# 网络搜索工具
search = [
    "tavily-python>=0.3.0",
]
# 图像生成工具
image = [
    "openai>=1.0.0",
    "replicate>=0.20.0",
    "pillow>=10.0.0",
]
# 完整安装
all = [
    "tavily-python>=0.3.0",
    "openai>=1.0.0",
    "replicate>=0.20.0",
    "pillow>=10.0.0",
]
# 开发依赖
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/agent_core"]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
